consolidate trait
