+++ crates/ferritin-plms/src/lib.rs
@@ -16,7 +16,6 @@ pub use esm2::esm2_runner::{ESM2Models, ESM2Runner};
 pub use featurize::StructureFeatures;
 pub use ligandmpnn::configs::ProteinMPNNConfig;
 pub use ligandmpnn::model::ProteinMPNN;
-pub use ligandmpnn::proteinfeatures::LMPNNFeatures;
 
 pub mod amplify;
 pub mod esm;
