+++ crates/ferritin-plms/src/ligandmpnn/mod.rs
@@ -1,16 +1,13 @@
-//! ferritin-amplify
+//! ferritin-ligandmpnn
 //!
 //! - utilities to convert sequence formats (PDB; mmcif) to ML-ready tensors.
 //! - CLI to handle the above.
 //!
 mod commands;
-
 pub mod configs;
 pub mod model;
 pub mod proteinfeatures;
 mod proteinfeaturesmodel;
 pub mod utilities;
-
 pub use configs::ProteinMPNNConfig;
 pub use model::ProteinMPNN;
-pub use proteinfeatures::LMPNNFeatures;
